+++ src/components/Header/NavDetails.tsx
